refactor(decompose): split webapp/e2e-tests/highest-value-system/electron-smoke-test.spec.ts into folder
